requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiortc==1.6.0
TTS==0.22.0
faster-whisper==0.10.0
google-generativeai==0.3.2
torch>=2.0.0
numpy>=1.24.0
av==10.0.0
python-multipart==0.0.6
websockets==12.0